requirements.txt
selenium>=4.21.0
faker>=25.0.0
pytest>=8.2.0
pytest-html>=4.1.1
python-dotenv>=1.0.1
